git-svn-id: http://wot-xvm.googlecode.com/svn@1799 802240b0-0b4d-18ce-4811-9324b01ec0ff

+++ trunk/addons/wn6&eff/xeff_to_eff.js
@@ -1,40 +1,49 @@
+// Скрипт для замены одного представления рейтинга на другое
+
 // Искомый текст и текст замены
 var text_sample1="{{xeff}}";
 var text_replace1="{{eff:4}}";
 var text_sample2="{{c:xeff}}";
 var text_replace2="{{c:eff}}";
 
-// Имя файла берём из аргумента или задаем XVM.xvmconf, если аргумент пуст
-if (WScript.Arguments.length<1) {
-    var file_name="XVM.xvmconf";
-}
-else {
-    file_name=WScript.Arguments(0);
-}
+var i=0;
+do {
+	// Имя файла берём из аргумента или задаем XVM.xvmconf, если аргумент пуст
+	if (WScript.Arguments.length<1) {
+		var file_name="XVM.xvmconf";
+	}
+	else {
+		file_name=WScript.Arguments(i);
+	}
 
-var fso=WScript.CreateObject("Scripting.FileSystemObject");
-// Переносим исходный файл во временный
-var file_name_tmp=file_name+".tmp";
-if(fso.FileExists(file_name_tmp))
-  fso.DeleteFile(file_name_tmp);
-fso.MoveFile(file_name,file_name_tmp);
+	var fso=WScript.CreateObject("Scripting.FileSystemObject");
+	if (!fso.FileExists(file_name))
+	  break;
+	// Переносим исходный файл во временный
+	var file_name_tmp=file_name+".tmp";
+	if(fso.FileExists(file_name_tmp))
+	  fso.DeleteFile(file_name_tmp);
+	fso.MoveFile(file_name,file_name_tmp);
 
-var fo=fso.OpenTextFile(file_name_tmp,1,false,false);
-var fr=fso.OpenTextFile(file_name,2,true,false);
+	var fo=fso.OpenTextFile(file_name_tmp,1,false,false);
+	var fr=fso.OpenTextFile(file_name,2,true,false);
 
-// Выполняем замену текста в каждой строке
-var re1=new RegExp(text_sample1);
-var re2=new RegExp(text_sample2);
+	// Выполняем замену текста в каждой строке
+	var re1=new RegExp(text_sample1);
+	var re2=new RegExp(text_sample2);
 
-while(!fo.AtEndOfStream){
-  var line=fo.ReadLine();
-  var line_replace=line.replace(re1,text_replace1);
-  line_replace=line_replace.replace(re1,text_replace1);
-  line_replace=line_replace.replace(re2,text_replace2);
-  line_replace=line_replace.replace(re2,text_replace2);
-  fr.WriteLine(line_replace);
-}
-fo.Close();
-fr.Close();
-// Удаляем исходный файл
-fso.DeleteFile(file_name_tmp);
+	while(!fo.AtEndOfStream){
+	  var line=fo.ReadLine();
+	  var line_replace=line.replace(re1,text_replace1);
+	  line_replace=line_replace.replace(re1,text_replace1);
+	  line_replace=line_replace.replace(re2,text_replace2);
+	  line_replace=line_replace.replace(re2,text_replace2);
+	  fr.WriteLine(line_replace);
+	}
+	fo.Close();
+	fr.Close();
+	// Удаляем исходный файл
+	fso.DeleteFile(file_name_tmp);
+
+	i++
+} while (i<WScript.Arguments.length);
